requirements.txt
# Existing Django packages
Django==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt
drf-spectacular

# NEW: Lambda adapter
awsgi2==1.0.2

# NEW: PostgreSQL driver for Lambda
psycopg2-binary

# NEW: Environment variable management
python-decouple

# NEW: CORS support for React frontend
django-cors-headers==4.3.1

# NEW: S3 storage for static/media files
django-storages
boto3